requirements.txt
streamlit
crewai-tools
litellm
python-dotenv
SpeechRecognition
crewai
solana==0.35.0
solders>=0.21.0,<0.22.0
qrcode
requests
